Keep it DRY
